backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

# Google Sheets (optional — only needed when DATABASE_BACKEND=sheets)
gspread==6.1.4
google-auth==2.37.0

# Dev / testing
httpx==0.28.1        # for FastAPI TestClient
pytest==8.3.4
pytest-asyncio==0.25.0
